Validacoes e criptografia adcionados

+++ Api/Api-CodeReview/Api-CodeReview/Service/BranchService.cs
@@ -1,6 +1,7 @@
 ﻿using Api_CodeReview.Context;
 using Api_CodeReview.Models;
 using Api_CodeReview.Repository;
+using Api_CodeReview.Repository.Interfaces;
 using LibGit2Sharp;
 using Microsoft.EntityFrameworkCore;
 using System;
@@ -15,10 +16,14 @@ namespace Api_CodeReview.Service
     public class BranchService
     {
         private readonly IBranchRepository repository;
+        private readonly IRepositorioRepository repositorioRepository;
+        private readonly ICommitRepository commitRepository;
 
         public BranchService(AppDbContext context)
         {
             repository = new BranchRepository(context);
+            repositorioRepository = new RepositorioRepository(context);
+            commitRepository = new CommitRepository(context);
         }
     
         public async Task<IEnumerable<Models.Branch>> GetBranches()
@@ -39,43 +44,64 @@ namespace Api_CodeReview.Service
         public async Task PutBranch(Models.Branch branch, int id)
         {
             if (id != branch.Id_branch)
-            {
-                return;
-            }
+                throw new Exception("Id's diferentes");
 
-            try
+            var repositorio = await repositorioRepository.GetById(branch.Id_repositorio);
+            if (repositorio == null)
+                throw new Exception("Repositorio não existe");
+
+            foreach (var item in repository.ListarPossiveisBranchs(repositorio))
             {
-                await repository.Update(branch);
-                await repository.Save();
-            }
-            catch (DbUpdateConcurrencyException)
-            {
-                if (!repository.BranchExist(id))
+                if(item.EndsWith(branch.Nm_branch))
                 {
+                    branch.Nm_email_dev = CriptografiaService.Encrypt(branch.Nm_email_dev);
+                    branch.Nm_email_review = CriptografiaService.Encrypt(branch.Nm_email_review);
+
+                    await repository.Update(branch);
                     return;
                 }
-                else
-                {
-                    throw;
-                }
             }
+
+            throw new Exception("Erro ao alterar dados da branch");
+            
         }
     
         public async Task PostBranch(Models.Branch branch)
         {
-            await repository.Post(branch);
+            if (repository.BranchExist(branch.Id_branch))
+                throw new Exception("Branch id ja existe");
+
+            var repositorio = await repositorioRepository.GetById(branch.Id_repositorio);
+            if (repositorio == null)
+                throw new Exception("Repositorio não existe");
+
+            foreach (var item in repository.ListarPossiveisBranchs(repositorio))
+            {
+                if (item.EndsWith(branch.Nm_branch))
+                {
+                    branch.Nm_email_dev = CriptografiaService.Encrypt(branch.Nm_email_dev);
+                    branch.Nm_email_review = CriptografiaService.Encrypt(branch.Nm_email_review);
+
+                    await repository.Post(branch);
+                    return;
+                }
+
+            }
+
+
+            throw new Exception("Erro ao alterar dados da branch");
         }
 
         public async Task DeleteBranch(int id)
         {
-            try
-            {
-                await repository.Delete(id);
-            }
-            catch (Exception)
-            {
-                throw;
-            }
+            if (!repository.BranchExist(id))
+                throw new Exception("id não existe ja existe");
+
+            if (commitRepository.CommitExistByIdBranch(id))
+                throw new Exception("essa branch ainda tem ultimo commit registrado");
+
+
+            await repository.Delete(id);
         }
     }
 }
